Adding Nidavellir, land of dwarfs...
Nidavellir is the main placeholder for all the libraries and executables targeting embedded platform(s).
This commit introduces one such example - Hreidmar.
This is a project that is based on STM32F4 demo sub-project that is provided with ChibiOS. The current version:
  - initializes the OS,
  - initializes the serialization engined over the serial device SD2,
  - does a simple echo of incoming ping-packets...

Also committing a not-very-perfect Makefile build system that is based on the ChibiOS build system (and includes some of the internal Makefiles)...